cpu util 8

+++ metrics/processors/pod_based_enricher.go
@@ -185,8 +185,8 @@ func updateContainerResourcesAndLimits(metricSet *core.MetricSet, container kube
 
         // populate CPU utilization (in percentage of limit or request)
         cpuUsageRate := metricSet.MetricValues[core.MetricCpuUsageRate.Name].IntValue
-        cpuLimit := metricSet.MetricValues[core.MetricCpuLimit.Name].IntValue
-        cpuRequest := metricSet.MetricValues[core.MetricCpuRequest.Name].IntValue
+        cpuLimit := float32(metricSet.MetricValues[core.MetricCpuLimit.Name].IntValue)
+        cpuRequest := float32(metricSet.MetricValues[core.MetricCpuRequest.Name].IntValue)
         cpuLimit = math.Max(cpuLimit, cpuRequest)
 
         // Default assumption is that there is no Limit or Request set on CPU
